pyproject.toml
[project]
name = "aria-framework"
version = "0.1.0"
description = "Artificial Intelligence Regulation Interface & Agreements"
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
authors = [
    { name = "ARIA Team" }
]
keywords = ["ai", "policy", "regulation", "framework", "governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",           # For YAML policy file handling
    "click>=8.0.0",          # For CLI interface
    "rich>=12.0.0",          # For rich terminal output
    "jsonschema>=4.0.0",     # For policy validation
    "jinja2>=3.0.0",         # For template rendering
    "typing-extensions>=4.0.0",  # For enhanced type hints
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=5.0.0",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/aria-framework/aria"
Documentation = "https://aria-framework.readthedocs.io"
Repository = "https://github.com/aria-framework/aria.git"
Issues = "https://github.com/aria-framework/aria/issues"

[project.scripts]
aria = "aria.cli:main"

[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aria"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=aria"
testpaths = ["tests"]